host_agent/pyproject.toml
# This file defines the workspace for the a2a-friend-scheduling project.
# It tells uv that this project is a monorepo containing multiple packages.

[project]
name = "a2a-friend-scheduling"
version = "0.1.0"
description = "A multi-agent system for scheduling pickleball games."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Shared ADK & A2A Dependencies
    "a2a-sdk>=0.3.6",
    "google-adk>=1.14.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv",
    "click",
    "uvicorn",
    "google-generativeai",
    "httpx",
    
    # FastAPI dependencies for /chats API
    "fastapi>=0.104.0",
    "pydantic>=2.0.0",
    
    # Database dependencies
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    
    # Firebase dependencies
    "firebase-admin>=6.0.0",

    # AWS dependencies
    "boto3>=1.34.0",

    # PDF processing dependencies
    "PyMuPDF>=1.24.9",

    # Kaitlyn's agent dependencies (future)
    # "langgraph"
]
